backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "图书管理系统后端 API"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
